feat: 히스토리 상세 화면으로 이동 및 히스토리 상세 조회

+++ data/src/main/java/com/starFaceFinder/data/source/local/FaceInfoDao.kt
@@ -23,7 +23,7 @@ interface FaceInfoDao {
                 " INNER JOIN SimilarFace ON SimilarFace.fid = FaceInfo.fid" +
                 " WHERE FaceInfo.fid = :fid"
     )
-    fun getFaceHistory(fid: Int): Map<FaceInfo, List<SimilarFace>>
+    fun getFaceHistory(fid: Long): Map<FaceInfo, List<SimilarFace>>
 
     @Insert
     fun insertFaceInfo(faceInfo: FaceInfo): Long
